Split types into own files
Makes it easier to maintain multiple types.
