+++ apis/postgresql/v1alpha1/database_types.go
@@ -22,84 +22,6 @@ import (
 	runtimev1alpha1 "github.com/crossplane/crossplane-runtime/apis/core/v1alpha1"
 )
 
-// A ProviderConfigSpec defines the desired state of a ProviderConfig.
-type ProviderConfigSpec struct {
-	// Credentials required to authenticate to this provider.
-	Credentials ProviderCredentials `json:"credentials"`
-}
-
-const (
-	// CredentialsSourcePostgreSQLConnectionSecret indicates that a provider
-	// should acquire credentials from a connection secret written by a managed
-	// resource that represents a PostgreSQL server.
-	CredentialsSourcePostgreSQLConnectionSecret runtimev1alpha1.CredentialsSource = "PostgreSQLConnectionSecret"
-)
-
-// ProviderCredentials required to authenticate.
-type ProviderCredentials struct {
-	// Source of the provider credentials.
-	// +kubebuilder:validation:Enum=PostgreSQLConnectionSecret
-	Source runtimev1alpha1.CredentialsSource `json:"source"`
-
-	// A CredentialsSecretRef is a reference to a PostgreSQL connection secret
-	// that contains the credentials that must be used to connect to the
-	// provider. +optional
-	ConnectionSecretRef *runtimev1alpha1.SecretReference `json:"connectionSecretRef,omitempty"`
-}
-
-// A ProviderConfigStatus reflects the observed state of a ProviderConfig.
-type ProviderConfigStatus struct {
-	runtimev1alpha1.ProviderConfigStatus `json:",inline"`
-}
-
-// +kubebuilder:object:root=true
-
-// A ProviderConfig configures a Template provider.
-// +kubebuilder:subresource:status
-// +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
-// +kubebuilder:printcolumn:name="SECRET-NAME",type="string",JSONPath=".spec.credentialsSecretRef.name",priority=1
-// +kubebuilder:resource:scope=Cluster
-type ProviderConfig struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Spec   ProviderConfigSpec   `json:"spec"`
-	Status ProviderConfigStatus `json:"status,omitempty"`
-}
-
-// +kubebuilder:object:root=true
-
-// ProviderConfigList contains a list of ProviderConfig.
-type ProviderConfigList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []ProviderConfig `json:"items"`
-}
-
-// +kubebuilder:object:root=true
-
-// A ProviderConfigUsage indicates that a resource is using a ProviderConfig.
-// +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
-// +kubebuilder:printcolumn:name="CONFIG-NAME",type="string",JSONPath=".providerConfigRef.name"
-// +kubebuilder:printcolumn:name="RESOURCE-KIND",type="string",JSONPath=".resourceRef.kind"
-// +kubebuilder:printcolumn:name="RESOURCE-NAME",type="string",JSONPath=".resourceRef.name"
-// +kubebuilder:resource:scope=Cluster
-type ProviderConfigUsage struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	runtimev1alpha1.ProviderConfigUsage `json:",inline"`
-}
-
-// +kubebuilder:object:root=true
-
-// ProviderConfigUsageList contains a list of ProviderConfigUsage
-type ProviderConfigUsageList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []ProviderConfigUsage `json:"items"`
-}
-
 // DatabaseParameters are the configurable fields of a Database.
 type DatabaseParameters struct {
 	// The role name of the user who will own the new database, or DEFAULT to
